+++ packages/frontend/src/hooks/dashboard/useDashboardChartReadyQuery.ts
